chore: make hook for product table
